+++ SaveToPdf.java
@@ -1,7 +1,0 @@
-import javax.swing.*;
-
-public class SaveToPdf {
-    public void SaveToPdf(JTextArea txtArea, String documentName){
-
-    }
-}