+++ CHANGELOG.md
@@ -1,7 +1,34 @@
 # FNotepad Changelog 
 #### (not completed jet)
 
-## FNotepad [v1.3.3][0] (2021-02-18)
+## FNotepad [v1.3.5][0] (2021-03-13)
+
+### Added
+- Italian FNotepad
+- Help files in all languages
+- Open shell from FNotepad
+- Add CONTRIBUTING file
+
+### Changed
+- README
+
+## FNotepad [v1.3.4][10] (2021-02-23)
+
+### Added
+- Help file
+- New Menu buttons
+- Added test rpm
+- Added windows installer
+- Logo
+
+### Changed
+- Fixed big counting bug
+- Menu
+- License, from MIT to GNU GPLv3
+- Removed unnecessary code
+
+
+## FNotepad [v1.3.3][9] (2021-02-18)
 
 ### Added
 - Setting for tabulator size in page settings (beta)
@@ -48,6 +75,8 @@
 
 ## FNotepad [v1.2.1][5]
 
+### Changed
+- Small bug fixes
 
 ## FNotepad [v1.2.0][4] (2021-02-09)
 
@@ -60,14 +89,29 @@
 
 
 ## FNotepad [v1.1.1][3]
+
+### Changed
+- Small bug fixes
+
+
 ## FNotepad [v1.1.0][2]
+
+### Changed
+- Bug fixes
+
+
 ## FNotepad [v1.0.0][1]
+
+### Added
+- First release of FNotepad
 
 
 [0]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/latest
-[8]:https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.3.2
+[10]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.3.4
+[9]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.3.3
+[8]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.3.2
 [7]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.3.1
-[6]:  https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.2.2
+[6]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.2.2
 [5]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.2.1
 [4]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.2.0
 [3]: https://github.com/fantastic-octo-garbanzo/FNotepad/tree/main/release/v1.1.1